add some more tests, fix a failure path from evil input in federateCreation functions, [codecov]
